Moving controllers into control
